[Lit] Refactor SpecialFunction table a expr mapping, NFC. (#6049)
This expanse SpecialFunctions.def to keep track of which expression
a given special function corresponds to, and expands it to include
a bit that indicates whether it is a reversed method.

This allows us to eliminate the switch statement in
getOpSpecialFunctions,
and expand it to support mapping reversed special functions.

While here, regularize the naming of various enums to be more
consistent.  We have precedence levels, tokens, special functions,
and expr nodes that are all similar but different naming.  This gets
us a step closer to being consistent.

MODULAR_ORIG_COMMIT_REV_ID: 73ed0b1f012c644a801419b443069c3553cf1777