[Lit] Rename and narrow declForSymbol mapping, NFC. (#6066)
This mapping is used to resolve ASTType's back to the ASTDecl for
that represents the underlying type.  This is therefore only used
for type-like things, so we don't have to keep track of it for
variables and functions.

MODULAR_ORIG_COMMIT_REV_ID: bed73a0df99a3107c23dc6c2c38c2841f505a301