[Lit] Change when symbols get applied to LIT::FuncOp's, NFC (#6067)
This changes the symbol application time to be after type checking,
which is step #1 of Issue #5854.

Unfortunately, this required a bunch of refactoring to the decorator
processing logic.  I was hoping to avoid this, but the exported
decorator needs the resolved symbol so it needs to run after signature
checking should take place.  This is really where all decorator
application should happen, so I'll continue to work in this.

MODULAR_ORIG_COMMIT_REV_ID: 24df78a6db9dcfb0d0cc4eef15aad91749b2f4ac