[Lit] Rearrange handling of decorators, and standardize 2-phase parsing. (#6072)
This eliminates the extra decorator cursor in ASTDecl in favor of
always starting a lazily parsed declaration with the start of the
declaration ... which includes the decorators.  This makes it easy
to expose the decorator list to the signature resolution phase of
that declaration, simplifying things.

MODULAR_ORIG_COMMIT_REV_ID: 1290e1ee525c07ae7c45974ff61719c81ad263e0