[Lit] Fix the @implements decorator to perform name lookup. (#6076)
This changes the implementation of this decorator to perform a proper
lookup of the referenced symbol in question, which will allow it to
continue to work when we start name mangling symbols.

The decorator logic could be further generalized IMO, to evaluate
arguments as MValues pervasively.  We can do this later when we have
more decorators.  Until then the architecture is closer to right.

MODULAR_ORIG_COMMIT_REV_ID: 350ff13894bf7d5bd1bcd3036db9a076c7800451