[LIT] Add initial Bool type for lit-stdlib (#6052)
This adds a Bool type (backed by pop.scalar<bool>) to lit-stdlib. This
is the start of a stdlib and will be expanded in the future. The
organization of the stdlib will also be figured out in the future.

MODULAR_ORIG_COMMIT_REV_ID: 21ee58f2b5a28d52885031eff275edc1b1a4dbf2