[KGEN] Emit C wrappers for exported KGEN functions (#6118)
We were previously rewriting the signatures of top-level kernels
in-place to make them compatible with C calling conventions for certain
types. This meant that we could only export functions that are not
called by any other functions (within a module). This was fine when we
compiled each kernel into its own object file where it was guaranteed to
be the top-level symbol, but this changed a while ago. This meant that
if you exported symbols `@a` and `@b`, compilation would actually fail.

This PR makes LowerKGENToLLVM generate C wrapper functions suffixed with
`_c` instead, and simplifies the handling of exported symbols. We drop
the `c-call` flag that was only used for testing and instead just
generate a C wrapper for every exported function.

Overall, this simplifies things a lot and lets us actually compile
complex modules.

MODULAR_ORIG_COMMIT_REV_ID: 22ae871c7e4f4a9afc5efc723d121637d2a6c6b8