[Lit] Diagnose ambiguous method overloads as redefinitions. (#6141)
This changes the parser to maintain an MLIR SymbolTableCollection for
the module it is building and check it when installing new functions
into it for conflict.  Given we mangle names to make them unique
according to overloading, we can use this as a simple way to detect
redefinition errors.

Note that this requires some pretty weird MLIR gymnastics.  MLIR
really isn't set up for operations that get a symbol figured out
after the operation is constructed.  We can make this work with enough
dodging and weaving though.

MODULAR_ORIG_COMMIT_REV_ID: cda8ce8b65931365236b772871753c93766d7d86