[Lit] Expand LookupResult to represent an overload set. (#6149)
This expands name lookup to be able to represent a lookup that finds
multiple declarations and puts them in LookupResult, along with the
relevant plumbing for that.

This still does not allow disambiguating a reference to one or another
of the candidates, but this gets us a lot closer to being able to
support that.

MODULAR_ORIG_COMMIT_REV_ID: 066a2faf5ec2ca6d980fb7a03aca282db470b124