[KGEN] Reland #6121 and #6122: Elaborator region handling simplification (#6155)
With an additional patch that fixes the release build crash.

[KGEN] Elaborate regions like generators (#6121)

One problem is that `bind_signature(:() -> index fn)` will fail if `fn`
resolves to a region reference, since `bind_signature` does not support
region references (which are just StringAttr).

One way to resolve this issue is to make the Elaborator treat region
bodies like generators (materialized as `SymbolConstantAttr`, so that
`bind_signature` can process them). This has the benefit of greatly
simplifying the Elaborator by unifying the paths for dealing with
generator calls and region calls.

[KGEN] Remove RegionReturn from Elaborator (NFC) (#6122)

This machinery isn't needed anymore since regions are handled like any
generators. The stack is maintained using recursive calls to the
elaborator.

[KGEN] Remove caching of nested region uses

This was causing a bug where the address of a deallocated op is giving
to a newly allocated op, causing the incorrect uses to be found for that
op. This also wasn't going to work anyways, since the uses would need to
be remapped when region bodies are hosted to generators. I'm just going
to remove this for now.

MODULAR_ORIG_COMMIT_REV_ID: 3dbb75df41df9779bb242457d72f4ebf27b8dfa0