[Lit] Expand LookupResult to represent an overload set. (#6156)
This expands name lookup to be able to represent a lookup that finds
multiple declarations and puts them in LookupResult, along with the
relevant plumbing for that.

This still does not allow disambiguating a reference to one or another
of the candidates, but this gets us a lot closer to being able to
support that.

This is a reapply of PR#6149 with some fixes.

MODULAR_ORIG_COMMIT_REV_ID: 1e25d50eca58621a584f31af022bff7a7ceb4540