[Lit] Implement support for function overload resolution and diagnostics. (#6162)
This teaches function call resolution to filter the candidate list
according to the actual parameter and value argument specified at the
site of the call, diagnosing an error if none of the candidates are
viable or if multiple are viable and ambiguous.

This is still lacking support for implicit conversions in overload
resolution.

MODULAR_ORIG_COMMIT_REV_ID: b8da14aeccd8be7a433b2e17855f37c565540e4d