[Support] Introduce `SmartVariant` for generic use over `PointerUnion` (#6112)
`llvm::PointerUnion` can not be used generically by callers since it
only works when all of the types are pointer types. Introduce
`SmartVariant` which allows for generically working with a variant-like
type. Specifically, if all of the types are pointers (i.e. there are
bits we can steal to represent the nullable value), the underlying
storage is `PointerUnion`: otherwise, fall back to the less
space-efficient `std::variant`. `SmartVariant` provides a uniform API
for working with the set of types by providing the LLVM casting infra:
`isa`, `get`, `dyn_cast`, etc. along with a similar API of checking for
nullability via `isNull()` on the `SmartVariant` itself.

Note:
- `dyn_cast` semantics when you fall back to use of `std::variant` seem
a bit weird, so open to thoughts there.
- doesn't work yet for llvm's `PointerLike` types: i.e. those that define a
specialization of `llvm::PointerLikeTypeTraits`.  In that case, we want to
treat those as pointerlike and use `PointerUnion` still.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 6a8d869392d8860588155421242bd0ac9f9767ec