[Lit] Suck a bunch of name lookup stuff into CallableValue ctor. (#6229)
This pulls the common pattern of "look up a named decl on a type and
produce a CallableValue for the overload set" into a CallableValue
constructor.  This simplifies a bunch of clients, and keeps the code
more consistent.

This is NFC other than a trivial change to the diagnostics to make
them more consistent.

MODULAR_ORIG_COMMIT_REV_ID: 04933fb982dd90a4bbe106aba3ad373c8f44e2bd