[Lit] When emitting conditions, elide __bool__ in common cases. (#6239)
It is very typical to do 'if' on the result of something that is already
a Bool type. In that case the __bool__ call is redundant and just wastes
compile time. Using the new CallableValue infra, we can easily skip
this.

MODULAR_ORIG_COMMIT_REV_ID: b16a780805b23553e3c017393c1bfb348f4e2690