[Lit] Simplify api to CallableValue ctor, NFC. (#6278)
There are two different modes of constructing this, the "emit an
error on failure" and "try looking up a thing" mode.  These are
distinct usecases, so split them and simplify the signature of the
ctors to reflect that.

MODULAR_ORIG_COMMIT_REV_ID: 1a9987be85a0682f68144824f02b3df1f024d7d1