[Tooling] Introduce common `lit.cfg.py` (#6299)
Our use of `lit.site.cfg.py.in` and `lit.cfg.py` is a lot of copy-paste
across the codebase. This makes refactoring/bug fixes in these files
cumbersome as we have so much duplication. This approach does not scale
well at all as we introduce more test directories.

To solve the duplication problem and mental/cognitive overhead when
developers copy-paste these files around, introduce a common
`lit.cfg.py` file. This provides sane defaults for setting up the lit
config appropriately and then individual test directories only need to
set a few properties on the `config` object itself.

The way this works is that in the top-level `CMakeLists.txt`, at CMake
configure time, we configure `utils/lit.common.configured.in` which
specifies these project-wide common config properties that are mostly
passthrough of CMake variables that get expanded and set on the config
object itself. This allows each current use of `lit.site.cfg.py.in` to
load this project-wide config to avoid mass repetition throughout the
codebase.

Apply this pattern in the `Support/test` directory only to keep this PR
small and as simple as possible while still exposing the core concepts
and abstractions. We can apply it everywhere else in a follow-up PR to
delete a bunch of code.

Partially addresses [Internal link] We
can close that issue once we remove all the duplication in a follow-up
PR.

MODULAR_ORIG_COMMIT_REV_ID: 95f57100edc696443faadc025d7d149c4f35d521