[KGEN] Scalarize variant codegen (#6223)
This PR changes the variant codegen to be "better" in the sense that it
no longer generates an alloca + lifetimes and a bitcast. Instead, it
generates the necessary ops to create and extract variant values in
registers. This means we don't need to depend on LLVM to run SROA, which
would not always happen. More importantly, this gives us a predtictable
way to create constant variants.

Fixes #3594

MODULAR_ORIG_COMMIT_REV_ID: 49963d2a7eb289c7b0b9f57c99c5e05ad2ff9b23