[KGEN/Support] Move TargetInfoAttr to MDialect (#6378)
Target abstraction will be important to clients other than KGEN. This PR
moves the TargetInfoAttr to MDialect and in KGEN introduces a
TargetParamAttr that just contains a TargetInfoAttr and always has a
TargetType -- it plugs the TargetInfoAttr into the parameter system.

This PR makes some subtle changes: `#kgen<target host>` is the new
syntax for the host attribute. This is to that the syntax for a concrete
target can still be `#kgen<triple = "etc", ...>`. I also pruned the
dependency on `llvm/MC/SubtargetFeature` since that class doesn't do
anything complicated and would require a dependency on LLVM MC which we
don't want.

MODULAR_ORIG_COMMIT_REV_ID: 323cbe26706b07464c674c358b24b014a7a9c5a5