[KGEN] LowerToObject: bring the LLVM pass pipeline logic from LLVM to our tree. (#6411)
We've been using the default LLVM optimization pass pipeline. Most
likely we don't need all the passes there and/or can benefit from
rearranging or adding other passes to it. This PR moves the pipeline
construction logic into our repo allowing us to tinker with it later.

The next planned step is to remove as many passes as possible without
regressing performance on kernels we currently generate.

Ultimately we also would like to rely on LLVM "magic" less and be more
explicit about optimizations we need in KGEN. E.g. vectorization and
unrolling would better be controlled explicitly in the kernel
implementation - this might not always be the case now, but we should
eventually address that. With this PR it would be easier to detect such
cases and get rid of the dependency on LLVM optimizations.

This the first step at resolving #4111.

MODULAR_ORIG_COMMIT_REV_ID: f477791f320599d6c069d30c0804d9c641959079