[Lit] Add test case for higher order parameter functions. (#6467)
With Jeff's work in PR6415 we are able to support passing functions
as parameters, which is the bedrock of "regions on calls".  The
major limiting capability is that we can't express the "not isolated
from above" use-case for generators.

MODULAR_ORIG_COMMIT_REV_ID: 330e1a38f4716653a9b8eee05f17109ed4b8776c