[KGEN/LIT] Move struct lowering to before elaboration (#6439)
This is a small code change with not much overall effects but it has big
ramifications for KGEN and the overall pipeline. This change effectively
turns KGEN struct decls and declref types into LIT concepts, which are
lowered away after LowerLIT, and before the KGEN elaborator. The reason
these types were kept around during elaboration is that they were
supposed to used to perform header-file generation on the
post-elaboration "target independent code". However, that idea is now
wrong because the elaboration in fact instantiates target-specific code,
since generators can be constrained on target-specific features.

Moreover, the structs were not actually being used to do anything. If
necessary, we can carry ABI metadata as function attributes to perform
header file generation post-elaboration, if that is what is decided. The
target-portable code is now at the LIT level.

Getting rid of structs pre-elaboration leads to a lot of simplifications
in the code, because properties of a declref type cannot be deduced
without a symbol table.

MODULAR_ORIG_COMMIT_REV_ID: 4fd410bb72f0f4cfb9ae3b6b4163bcbabf6f9389