[KGEN] Add support for index dtype to SIMDAttr and folders (#6494)
This PR adds support for creating SIMD constants of index dtype. Values
of index dtype are always stored as 64-bit integers and are treated as
signed. This PR also expands the arithmetic POP op folders to support
index dtypes. The logic here follows the index dialect: if the result
computed with 64-bit truncated to 32 bits and 32-bit arithmetic are the same,
then the result is accepted.

We could implement the `interpret` hook for these ops and perform the
artihmetic knowing the target-specific size of `index`, but this will
do for now.

MODULAR_ORIG_COMMIT_REV_ID: 2f62059349addefb2ae3979919bf36b774cbe07a