[Support] Utility for making aligned unique pointers (#6529)
We use `alignedAlloc` and `alignedFree` in several places, but use it in
such a way where we lack a clear ownership model of the memory in some
cases. This is likely because it's a bit cumbersome/verbose to spell out
`std::unique_ptr<T>` with a custom deleter using our `alignedFree`
deleter for releasing the memory.

Add a utility: `makeAlignedUniquePtr` which does exactly that: given an
alignment and size, make a `unique_ptr<T, TDeleter>` where `TDeleter` is
fixed to be the deleter for `alignedFree`. Apply it in
`CompilerRT/Benchmark.cpp` as an example, but it can be applied very
easily in nearly all cases of `alignedFree` to use unique ownership.

MODULAR_ORIG_COMMIT_REV_ID: 3abc2adb8444be6c1a5722545ae1cbdb309824c4