[Kernels] Add high-order map operation over buffers (#6546)
This implements a high-order functional map operation over buffers. The
`map` operation takes a function and a buffer and applies the function
to each element of the buffer. The function is expected to be a `f32`
due to a limitation in the current lit parser, but that can easily be
changed once [Internal link] is fixed.

With this as a template, we can implement all the fancy transforms
that we have defined in the MLIR code.

BUFFER_ORIG_REV_ID: affcf690865bf8d01380214412e542f56a3a9ba3