[Lit] Reject function calls missing parameters. (#6627)
We allow forming references to symbols with no parameters bound
so they can later be bound, but need them to be bound in function
call position.  We missed this, which leads to incorrectly accepting
bad code and crashing later in the compiler.

This fixes Issue #6624

MODULAR_ORIG_COMMIT_REV_ID: a656e0f626797737590d07297bfc2463896666e6