[Support] Install vscode-lit by linking it to the extension directory (#6551)
Previously we were installing the vscode-lit extension by building it.
This requires npm which is a bit annoying to install. There is another
method which is to just side-load the extension by placing it in the
extensions directory.

This also moves the code for linking directories out to the common
python packages. This code is shared between build.py and vscode-init
and is a bit complicated because it has to deal with some os
intricacies.

MODULAR_ORIG_COMMIT_REV_ID: 781be84df9c57357089576b951415b7ea8942a75