[Lit] Introduce a first class IR representation for alias forward decls. (#6653)
`alias x: Int` is a forward declaration of a parameter that will be
defined as the result parameter in a function call.  Add an MLIR op
to represent this, and diagnose when they are not actually defined
(right now, never - because that isn't wired up yet).

MODULAR_ORIG_COMMIT_REV_ID: cc437a2746a6297b17fe538675d4a35678dc091a