[KGEN] Add KGENDialectInlinerInterface. (#6695)
We need to be able to inline kgen.func ops that have the force_inline attribute, so this patch adds the correct dialect interface. For now, we ignore anything that's not a kgen.func to avoid having to deal with things like parameters.

MODULAR_ORIG_COMMIT_REV_ID: 5b295454e8106478c0062657c287a92adbf4ee9a