[LIT] Wrap the input file in a module (#6708)
This greatly simplifies quite a few things in the model,
and makes parsing/tooling/etc significantly easier to
work with. Various tests are updated to account for the
module prefix, and the definition of ErrorOr was updated
to not hardcode the symbol reference to Error.

See #6370

MODULAR_ORIG_COMMIT_REV_ID: 613777a5015317b3b1c24da3c5a46e3094bcdfe4