[GenericML] Prepare BufferRef for supporting 'foreign buffers'. (#6629)
When crossing between framework- and modular-ops we convert between
framework- and modular-tensors. For TFL we managed to do that mostly
without any buffer copying, however it worked only because our TFL
lowering mimics TFL's caching strategy. (See the comment in
TFLite/include/TFLite/TFLTensor.h for how that works.)

For TF we're out of luck using the TFL approach. Using 1x5
bert-base-uncased as an example, and disabling buffer sharing in TFL as
an estimator for the possible impact on TF once we support a similar set
of modular ops, it turns out we copy 330MB taking 350ms (!). So baring
great op coverage it seems worthwhile to avoid the copies.

This PR just does some prep work so the main change (#6630) is
managable:
- BufferRef reworked a little bit so that the 'native' buffer fields are
collected. We drop isOwned() in favor of getOwnership() since we're
about to get two notions of 'owned'.
 - Both BufferRef and TFPTensor will need support for releasing.
- The tfp.m_tensor_to_tf_tensor and tfp.tf_tensor_to_mtensor casts are
reworked so that we can layer in the changes exploiting 'foreign'
buffers as a simple local change. They'll need access to the tfp.context
for that.
 - Got rid of Tensor::create since the ctor is just fine.
- A small refactor in MTFOps.cpp so any lowering implementation can get
at the tfp.context.
- Make sure the TFExecutorContextRef needs to be copied into addTask
closures. My later changes were triggering asan use-after-free as is.
- Let BufferRef and Tensor copy from an existing buffer since was
getting sick of all the memcpys.

MODULAR_ORIG_COMMIT_REV_ID: 7a2a97a6e0bbdbcade482bcfa750de5c7b91323c