[LLCL] Avoid unnecessary extra ref to AsyncValue for andThenAsync (#6768)
We're starting to rely on v.getPointer()->isUnique() to test when we
hold the unique AsyncValueRef<T> v. In #5912 we introduced andThenAsync
so that BEFExecutor::processUsedByRecords would give an addTask closure
the chance to finish and release extra AnyAsyncValueRef captured in its
closure before downstream primitives are scheduled. Unfortunately the
andThenAsync itself captured a copy of the very reference we were trying
not to hold onto. This PR distinguishes the 'void' vs
'AnyAsyncValueRef&' waiter flavors to avoid that capture. In this way
the reference count for SSA values in the BEF graph matches the true SSA
value usage count, and the isUnique runtime optimizations fire as
expected.

MODULAR_ORIG_COMMIT_REV_ID: b29fd18b5cbdf2084c73e81e42daaab1bf4d5a79