[Lit] Build more explicit syntax info into call emission, NFC. (#6787)
Instead of passing the expr into the call emission, pass an enum
that specifies why the call is being done.  The problem this solves
(vs passing the expr) is that implicit conversions may be done for
statements (not exprs) and we can't pass the statement in as an expr.
Passing the expression to be converted as the call expression is
wrong and can lead to confusing diagnostics.  This directly models
what we need.

MODULAR_ORIG_COMMIT_REV_ID: 68226179813e3b1e366e3755fde4ddc4a298462f