[Python API] Allow testing with fake C API (#6860)
There is not an easy way to allow for testing the C++ code that sits
between the Python binding glue (PyBind11) and the underlying C APIs.
Enable the ability to test this layer (and not be coupled to the real C
API) by injecting a template parameter to the `InferenceSession` to
control behavior. Add a "real API" wrapper which delegates to the real C
API function and a "fake API" that allows for inspecting various
behavior desired but is still API-compatible with the real API. This
allows for testing various parts while the tests can be quick and not
coupled to the real API.

This will be particularly useful to not rely on the real C API for
loading and compiling models: this level of testing can use a fake model
path that is "static" and returns a constant value for the prediction,
for example.

Moreover, this is useful to test this layer so we can -- in parallel --
explore other binding frameworks, such as `nanobind` while being backed
by tests, and also implement the real C API without being blocked from
Python API development.

These tests require RTTI and `-fexceptions` since the binding framework
requires it. They also embed a Python interpreter and start it once when
running the tests, which is required before calling any Python code
under the covers.

MODULAR_ORIG_COMMIT_REV_ID: c000c7980620b4d3380608540313b4314dbec410