[Lit] Move check for alias forward defition completeness. (#6921)
This moves the check from the body resolver for alias to the body
resolver for the enclosing function.  The problem is that we'd
incorrectly reject things like this:

```
fn f():
  alias x: index
  let y = function[()->x]()
```

because the body resolver for 'x' would be run before the body
resolver for 'y', and we'd diagnose 'x' as never being defined.

Unfortunately moving it to the function body resolver doesn't work
with our current design either, because the function body resolver
isn't forcing the signature of x to be resolved before it is done.
Solving this will require some more work on decl resolution in
functions which is tracked by Issue #5975.  This is still a useful
step though because this error will be caught by kgen, and we won't
reject invalid functions.

MODULAR_ORIG_COMMIT_REV_ID: 117eab18dffa8afd189d0fe86e910888e458510d