[Cache] Fix a subtle bug causing memory corruption. (#6966)
The code was accidentally using the move assignment operation instead
of placement new to initialize a mapped_file_region in a union.  The
problem is that the element inside the union isn't initialized at all
so this would run the "unmap" operation as part of the move assignment
operation, unmapping a completely random pointer in your address space.

This leads to... bad things that are hard to debug.  I believe that this
is the root cause of Issue #6961.

Note: Pair debugged with Aman.
MODULAR_ORIG_COMMIT_REV_ID: 6b853026659c8607b8ab65fd7f7a6a178908f9d6