[Lit] Introduce a new `@nodebug_inline` decorator. (#6980)
Many of the things at the bottom level of the Lit stack are trivial
zero-abstraction wrappers around MLIR things, for example, the `+`
operator on Int or the `__bool__` method on Bool itself.  These
operators need to be force inlined even at -O0, but they have some
additional things that we need to wrestle with:

1) In no case would a user actually want to step into the __bool__
   method on Bool or the + method on Int.  This would be terrible
   debugger QoI for unless you're debugging Int itself.  We need
   something like `__always_inline__, __nodebug__` attributes that
   clang uses in headers like xmmintrin.h.

2) Similarly, these "operators" should be treated by users as
   primitives: they don't want to know about MLIR or internal
   implementation details of Int.

3) These trivial zero abstraction things should be eliminated early
   in the compiler pipeline so they don't slow down the compiler,
   bloating out the call graph with trivial leaves.  Such thing
   slows down the elaborator, interferes with basic MLIR things
   like fold(), bloats out the IR, or bloats out generated debug info.

4) In a parameter context, we want some of these things to get
   inlined so they can be simplified by the attribute logic and
   play more nicely with canonical types.  This is just a nice to
   have thing those of us who have to star at generated IR. :-)

The solution to this is a new `@nodebug_inline` decorator. This
decorator causes the parser to force-inline the callee instead of
generating a call to it.  While doing so, it gives the operations
the location of the call itself (that's the "nodebug" part) and
strips out let decls that were part of the internal
implementation details.

This is a super-power-user-feature intended for those building
the standard library itself, so it is intentionally limited in
power and scope: It can only be used on small functions, it
doesn't support regions, by-ref, throws, async, etc.

This patch only adopts it in one place in the standard library,
but subsequent patches can adopt it more pervasively.

MODULAR_ORIG_COMMIT_REV_ID: a494d04823a25de94c1c3933b55181e79604e0f1