[Lit] Optimize the statement parser a bit.  NFC (#7001)
We create a varDeclCursor to handle implicit declarations of variables
in 'def' context's, but there are lots of other contexts that we parse
statements into.  Rework the logic to only create (and destroy) this
operation for 'def's.

MODULAR_ORIG_COMMIT_REV_ID: c296076432f9819bb8c77a2a457662f092272bb7