[KGEN][QoI] Clean up the ExecutionEngine, NFC. (#6996)
The only user of the `add` signature that took a list of KGEN::FuncOp was in SelectFastestFunction, which means we could hoist that functionality up to the callsite and clena up the interface of the ExecutionEngine. It also removes the dependency of CompiledFunc on KGEN::FuncOp.

MODULAR_ORIG_COMMIT_REV_ID: bfcb0f1054635afb9fdd79a5185f0e00f773a5c9