[KGEN][QoI] Add `produceStandaloneObjectAttr`. (#6997)
This patch adds a utility function to the KGEN::ObjectCompiler that produces a standalone object as a DenseResource attribute that can be attached to another operation. The benefit to this approach is two-fold: (1) it keeps the object file out-of-line in the IR, and (2) it allows the client to call Cache::{deflate,inflate}Constant on the operation it's attached to.

MODULAR_ORIG_COMMIT_REV_ID: 9c94bac033c2c4862da613bdf77dad09b4bfff5e