[KGEN][QoI] Avoid copying all data into the ORCJIT. (#6998)
The reason we copied the data was because we needed to ensure the data wasn't destroyed before the JIT - the beauty of ref-counted buffers is that we don't have to do that! We can instead tie the lifetime off the buffers to the lifetime of the JIT itself.

We should still work on #6989 - we shouldn't be materializing these buffers until we are ready to link them, but for now this is a big improvement.

MODULAR_ORIG_COMMIT_REV_ID: 341818a1e6b66e9fa220105f3069b47785bc8ffa