[Lit] Implement support for parsing/type checking * and / markers. (#7023)
Add support for parsing *args as a new VarArg convention.  Not
much is done with this yet.  While here add support for */ markers.

The bare '*' and '/' markers in an argument list have nothing to
do with variadics it turns out.  They are markers that change the
interpretation of keyword arguments.  Parse these and check them,
and add the result to ParsedArgument, even though we're not doing
much with keywords for the forseeable future.

MODULAR_ORIG_COMMIT_REV_ID: f2396eb3da1cdecce16f1285e4f4046947be66ad