[KGEN/Support/LLCL] Update AsyncSideEffectMap::await to return an error. (#7048)
AsyncValues can hold errors, and the thing you're waiting on may have error'd. When you call await, you will likely want to know if an error occurred, and handle it at that time. This returns the error that might be on the AsyncValue, and modifies the callsites to propagate/handle that error.

MODULAR_ORIG_COMMIT_REV_ID: 80dfadf21bc745d15ca5c24157452546807042ee