[Kernels] Introduce a single op which calls into Intel AMX (#6938)
This introduces the tdpbssd which is a lit functions that invokes the
AMX instructions.

The op compute dot-product of bytes in tiles with a source/destination
accumulator. Multiply groups of 4 adjacent pairs of signed 8-bit
integers in a with corresponding signed 8-bit integers in b, producing 4
intermediate 32-bit results. Sum these 4 results with the corresponding
32-bit integer in dst, and store the 32-bit result back to tile dst.

LINALG_ORIG_REV_ID: 19bc086323e449706132779abe1d492ad5b56a33