[Kernels] Fix types of the kgen.signature in a bunch of places (#7097)
This fixes the input and output parameter types of kgen.signature in a
bunch of places. These were binding the type to the outer scope which
was not correct and was causing a bunch of weird issues.

ALGORITHM_ORIG_REV_ID: 65b3b7514cd11212f47cb6e0092ba5c4aa82de2c