[KGEN] Add `pop.atomic.cmpxchng` operation to atomicaly modify memory (#7105)
The `pop.atomic.cmpxchg` operation is used to atomically modify memory.
It loads a value in memory and compares it to a given value. If they are
equal, it tries to store a new value into the memory.

There are three arguments to the `atomic.cmpxchg` instruction are:

  - an address to operate on
  - a value to compare to the value currently be at that address
- a new value to place at that address if the compared values are equal.

The type of `cmp` must be an integer or pointer type whose bit width is
a power of two greater than or equal to eight and less than or equal to
a target-specific size limit. `cmp` and `val` must have the same type,
and the type of `pointer` must be a pointer to that type.

The success and failure ordering arguments specify how this cmpxchg
synchronizes with other atomic operations. Both ordering parameters must
be at least monotonic, the failure ordering cannot be either release or
acq_rel.

The return type is a KGEN struct type with any integer or pointer
followed by a boolean value.

Example:

```mlir
%b0 = pop.atomic.cmpxchg %ptr, %cmp, %val monotonic monotonic :
    !pop.pointer<scalar<index>>
```

This is the first half of
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: d82afca37722a30d44b5408b8767f4adc438cc90