[LIT] Allow adding regions to `__mlir_op`'s (#7126)
This PR implements a super-duper-mega hack to allow operations to
implement regions. MLIR operations specified with `__mlir_op` can take
an optional `_region` attribute that lists references to the region
bodies (max 1 region right now due to lack of list `[]` literal), but it
uses a string attribute to lookup the function decl in the nearest scope
and forces the body to be resolved. It then moves the body into the
operation and deletes the terminator, since otherwise the ops in the
body may not verify (if they expect a certain parent op).

The way to clean this up is to add a special `mlir_region loop_body`
declaration (and associated LIT dialect operation) and maybe a
`#lit.mlir_region_ref` attribute to provide stronger invariants.

MODULAR_ORIG_COMMIT_REV_ID: 0db094af04ab603301788fbb00943f94ab0201d0