[KGEN] Minimize the elaborator interface, NFC. (#7164)
This patch prepares for adding a new elaborator implementation by setting up the elaborator as a minimal interface that the IREvaluator can use without needing to call into all the methods the elaborator can provide.

This means that we again hide most of the methods on the elaborator interface inside an ElaboratorImpl in the .cpp file, which makes changes a *little* harder but the encapsulation is valuable for ensuring the coupling between the IREvaluator and the Elaborator is as small as possible, given that they're dependent on each other.

MODULAR_ORIG_COMMIT_REV_ID: a5e23fbc493270fbf9c5df46dde1926a3a116bab