[Kernels] Switches the high-level map function to take a closure (#7187)
This removes the unary / binary map functions and instead provides a
single interface to map that takes a closure. It's up to the passed in
function to perform the loads / stores given an index.

This is the first part of transititioning all the generators to take
closures.

All the heavy lifting was done by @Mogball to all for this.

ALGORITHM_ORIG_REV_ID: 0fd71fd8da8e444bb1930f5b1697dea068f00a6c