[KGEN] Give the SIMDAttr assembly format splatness (#7249)
This was mentioned in a KGEN sync at some point, but this PR adds splat
syntax for the SIMDAttr in both generic and parameter form. The
splatness is just in the syntax sugar -- when a splat value is parsed,
it is replicated according to the vector size.

MODULAR_ORIG_COMMIT_REV_ID: 0d56a153f7d8482eafc2737518179d5a7232be44