[Lit] Provide a pretty struct initialization syntax, e.g. `Int{value=4}`. (#7334)
This uses the dictionary subscript syntax on types to cons up a value
of a struct type inline as an expression, which is super helpful for
implementing __new__ methods.  This implementation doesn't support
parameters, but I'll add that in the next patch.

For full disclosure, this is something that we're likely to want to
revisit in the future.  When ownership comes in, we'll probably move
everything from __new__ to an __init__ model which will eliminate the
need for this.  If we do that, I would like to make this syntax to be
overloadable/definable by types, enabling a c++ initializer_list sort
of syntax with an explicit type, e.g.: `DenseMap[key:1, key2: 2]`.

This this still going to be very handy in the interim, so it seems worth
doing until we have a nice replacement.

MODULAR_ORIG_COMMIT_REV_ID: 625e6a21d74f251a545e982a2567e095764c8979