[KGEN] Add an MLIROperatorAttr (#7323)
The `#kgen.param.mlir_op` attribute is a SignatureType parameter
attribute that represents an MLIR operation. It contains the name of the
operation and an attribute dictionary for it. The operand and result
types of the operation are stored in its signature type. The operation
can be parameterized on any of its input and result types or attributes
who are TypedAttrs. 'bind_signature' will bind parameters by placing
them inside the attribute dictionary and substituting them in the input
and result types.

This PR also teaches the 'apply' operator to fold the operation if
possible, in keeping with the "fold on construction if possible" spirit
of ParamOperatorAttr, however it may be preferable to do this in the
Parser or Interpreter, since constructing IR inside an attribute's
getter is undesirable.

MODULAR_ORIG_COMMIT_REV_ID: 1bc877a9524af09540aca575eeb76bac1ac7824a