[Lit] Improve error message in subscript syntax. (#7344)
It's a common error to use a = in a subscript or dictionary instead
of a :.  Tailor the error message to handle this nicely.  While we are
at it, build out our infrastructure for fixit hints and produce one
to rewrite to the correct form.  Before:

```
x.lit:8:48: error: expected ']' in call argument list
    return __mlir_op.`lit.struct.create`[value = 42]()
                                               ^
```

After:

```
x.lit:8:48: error: expected ':' in subscript slice, not '='
    return __mlir_op.`lit.struct.create`[value = 42]()
                                               ^
                                               :
```

This fixes #7307

MODULAR_ORIG_COMMIT_REV_ID: 37e573d2e3c1aaccd97b41372e0ad45f88bf7dfd