[LLCL] Support 'consuming' versions of andThen/emplace (#7256)
Part of #7069.

We're starting to rely on AsyncValue::isUnique to enable copy-vs-move
(and eventually copy-on-write) optimizations:
```
void my_consumer(AsyncValueRef<Foo> result) {
  if (result.getPointer()->isUnique()) {
    // Super cool trick exploiting that we hold
    // the only reference to the Foo in result.
  } else {
    // Oh well.
  }
}
```

However currently async primitives are written as:
```
AsyncValueRef<Foo> my_producer() {
  AsyncValueRef<Foo> result = ...
  addTask(runtime, [result = result.copy()]() {
    // Calculate a foo, at great expense.
    result.emplace(foo);
  }
  return result;
}
```
This means my_consumer may or may not see a stray additional reference
to the result, depending on a race for destroying the addTask closure.

We fix this by introducing a 'consuming' version of emplace:
```
AsyncValueRef<Foo> my_better_producer() {
  AsyncValueRef<Foo> result = ...
  addTask(runtime, [result = result.copy()]() {
    // Calculate a foo, at great expense.
    std::move(result).emplace(foo);
  }
  return result;
}
```
This version of emplace will keep the result reference alive only as
long as necessary, and will guarantee the reference will be destroyed
before any downstream waiter is triggered.

The ergonomics is not great: this forces users to understand move
semantics, move methods, mutable closures, and move capturing closures.
But eventually most of this machinery will be generated rather than hand
written so sense is we can tollerate this.

- The canonical AsyncValue Waiter is now arg-free. If the waiter needs
to keep the AsyncValue alive it must capture it within its closure.
- A 'consuming' version of andThen is added as convenience for that
case. Interestingly the vast majority of existing code was already using
the no-arg version.
- At least for the moment, force waiter types to be concrete, ie
AsyncWaiter::Waiter vs AsyncWaiter::ConsumingWaiter. Mostly to help my
brain not hurt so much, but also I was suspecting the compiler was being
too forgiving of function type mismatches. We can go back to the
templated form once the dust has settled.
- A 'consuming' version of emplace is now the main impl, with the
original method added as a helper. That will consume a (or the last)
reference to the AsyncValue and reset it just before the first waiter is
called.
- Beef up AsyncValueTest.cpp (was AndThenTest.cpp). Still pretty anemic
though.
- Change TFPrimitives.cpp fallback_execute to use new idiomatic style.
All the other async primitives however are still    pending :-|
- Left TODOs in a few places where the existing code was suspect.

MODULAR_ORIG_COMMIT_REV_ID: 860c38fa0c64d8168516fba6be11b5298a9d2b70