[KGEN] Add a `build_info` attribute and build_info_get_field POC (#7295)
This adds an initial `build_info` attribute that is understood by kgen.
The build info contains one field "type" which is the build type of the
modular codebase. We can use the build info to add other fields such as
version, whether search is enabled, etc.

The immediate use case for this is in lit where I want to be able to
write an assert statement which is conditioned on whether we are in
debug mode or not.

MODULAR_ORIG_COMMIT_REV_ID: 541b00d891b729d9efc575a54621c74f492cd694