[KGEN] Give result parameters names. (#7407)
This PR gives function result parameters names in KGEN and Lit. This is
primarily to support a use-case in MO/GraphCompiler, which requires
declaring and using result parameters in the signatures of graphs to use
them in the result types of the graph.

KGEN/LIT have uses cases for this as well: interfaces. The return type
of an interface can itself be parametric on the interface
implementation. This current PR is entirely cosmetic and implements the
wide-reaching code change to just give result parameters names.

In this next PR, the parameter verifier will be taught how to handle
result parameters, which requires some nuanced changes to use-def edges
in the parameter graph -- it will not longer be between operations that
define and use parameters, but the declaration attributes themselves.
And then in the next PR, the elaborator will be upgraded to support
named result parameters.

MODULAR_ORIG_COMMIT_REV_ID: a7205cb44e64b47bb34b2f75140054d04e403248