[KGEN] Rename `alloca` variable to `ptr`, NFC (#7466)
[KGEN] Rename `alloca` variable to `ptr`, NFC

In the code review for #7213 it was suggested to not use `alloca` as a variable name, since it may conflict with the Linux `alloca(3)` function. I modified the spot that was commented upon, but not this other variable name I introduced. This changes it to an innocuous name, `ptr`.

MODULAR_ORIG_COMMIT_REV_ID: c3ecde68c606dec978cf78da7ff70a9611f34df5