[LLCL] Export the profile level flag setting as a cmake flag (#7606)
This exports the profile level flag setting from llcl from being a
compile-time value the one has to change into a cmake flag that we can
pass in as a build argument. This was the intent from the very beginning
(thus resolving the todo comment). This will also allow us to use the
option in other parts of the code without brining in LLCL as a
dependency.

MODULAR_ORIG_COMMIT_REV_ID: c4a14ce249ac5984301f46fdcc409eb4cc4869a1